refactoring Tag tests
