+++ tests/unit/Tag/AbstractTagSetup.php
@@ -11,11 +11,11 @@
 
 declare(strict_types=1);
 
-namespace Phalcon\Tests\Fixtures\Helpers;
+namespace Phalcon\Tests\Unit\Tag;
 
 use Phalcon\Tag;
-use Phalcon\Tests\Fixtures\Traits\DiTrait;
 use Phalcon\Tests\AbstractUnitTestCase;
+use Phalcon\Tests\Fixtures\Traits\DiTrait;
 
 abstract class AbstractTagSetup extends AbstractUnitTestCase
 {
