+++ tests/unit/Tag/PreloadTest.php
@@ -16,7 +16,6 @@ namespace Phalcon\Tests\Unit\Tag;
 use Phalcon\Http\Response;
 use Phalcon\Tag;
 use Phalcon\Tests\AbstractUnitTestCase;
-use Phalcon\Tests\Fixtures\Helpers\AbstractTagSetup;
 use Phalcon\Tests\Fixtures\Traits\DiTrait;
 
 class PreloadTest extends AbstractUnitTestCase
