+++ tests/unit/Tag/SubmitButtonTest.php
@@ -13,8 +13,6 @@ declare(strict_types=1);
 
 namespace Phalcon\Tests\Unit\Tag;
 
-use Phalcon\Tests\Fixtures\Helpers\AbstractTagSetup;
-
 class SubmitButtonTest extends AbstractTagSetup
 {
     /**
