+++ tests/unit/Tag/TextAreaTest.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
 namespace Phalcon\Tests\Unit\Tag;
 
 use Phalcon\Tag;
-use Phalcon\Tests\Fixtures\Helpers\AbstractTagSetup;
 
 class TextAreaTest extends AbstractTagSetup
 {
